Adding new code

+++ mvc-app/src/main/java/com/tribepay/mvcapp/springdata/BankSpringDataRepository.java
@@ -3,5 +3,6 @@ package com.tribepay.mvcapp.springdata;
 import com.tribepay.mvcapp.entity.Bank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BankSpringDataRepository extends JpaRepository<Bank,Integer> {
+public interface BankSpringDataRepository extends JpaRepository<Bank,Long> {
+
 }
